+++ changelog.md
@@ -1,5 +1,10 @@
 # changelog
 
+## version 1.1.0
+- add new component radio-player
+- add types in $createElement
+- add new style retro card
+
 ## version 1.0.9
 - delete cjs & mjs for combined in one folder called core.
 - add $root in the core module.